% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseSchema.R
\name{addCohortDefinitionSet}
\alias{addCohortDefinitionSet}
\title{Insert atlas cohort definition set in to database}
\usage{
addCohortDefinitionSet(
  connection,
  config,
  cohortDefinitionSet,
  webApiUrl = NULL,
  exposure = FALSE
)
}
\arguments{
\item{connection}{DatabaseConnector::connection}

\item{config}{rewardb global config}

\item{cohortDefinitionSet}{A data.frame of cohorts containing -> cohortId, cohortName, json, sql fields}

\item{webApiUrl}{atlas source the definition set comes from (insert definitions from different sources sepratley)}

\item{exposure}{If exposure, cohort is treated as an exposure, drug domains are captured}
}
\description{
Creates individual references in Reward db for each cohort definition in the cohort definition set.
Note: Call this function separatley for exposure and outcome cohorts and different atlas sources.
}
