% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseSchema.R
\name{addSubsetDefinition}
\alias{addSubsetDefinition}
\title{Add subset definition}
\usage{
addSubsetDefinition(
  connection,
  config,
  cohortSubsetDefinition,
  targetCohortIds,
  exposure = TRUE
)
}
\arguments{
\item{connection}{a}

\item{config}{b}

\item{cohortSubsetDefinition}{subset}

\item{targetCohortIds}{target cohorts}

\item{exposure}{TRUE}
}
\description{
Adds subset definition to reward reference database
}
