% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-CalibrationPlotModule.R
\name{calibrationPlotServer}
\alias{calibrationPlotServer}
\title{Calibration plot siny server module}
\usage{
calibrationPlotServer(
  id,
  model,
  selectedCohort,
  dashboardControlSelector = "targetCohortId"
)
}
\arguments{
\item{id}{shiny namespace should be consistent with UI}

\item{model}{SccDataModel R6 class instance}

\item{selectedCohort}{cohort object reactive}
}
\description{
Display calibration plots for reward cohorts based on data available in system
Also shows Expected Absolute Systematic Error
}
