% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{copyCgTables}
\alias{copyCgTables}
\title{Copy Cohort Generator Tables to Dashboard Schema}
\usage{
copyCgTables(config = config::get(), dashboard)
}
\arguments{
\item{config}{Project configuration (`config::get()`).}

\item{dashboard}{Dashboard config object.}
}
\description{
Copies all required CohortGenerator results tables and concept set tables to the dashboard’s results schema, filtering cohorts as necessary.
}
\details{
Copies only results tables relevant to cohorts included in the dashboard.
}
\seealso{
[CohortGenerator::getResultsDataModelSpecifications()]
}
