% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{createDashboard}
\alias{createDashboard}
\title{Create And Upload a Full Dashboard Dataset}
\usage{
createDashboard(dashboardName, config = config::get())
}
\arguments{
\item{dashboardName}{Character. Name of dashboard (must exist in your config).}

\item{config}{Configuration list.}
}
\description{
Orchestrates the creation and population of all required datasets/tables for a dashboard from SCC output and meta information, then calibrates meta-analysis results.
}
\details{
Calls all required functions to generate analysis, upload results, copy tables, add meta info, and calibrate results.
}
\seealso{
[getDashboardConfig()], [createDashboardResults()], [uploadDashboardData()]
}
