% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardData.R
\name{createDashboardConfiguration}
\alias{createDashboardConfiguration}
\title{Create stand-alone dashboaard database from reward data}
\usage{
createDashboardConfiguration(dashboardConfigPath, overwrite = FALSE)
}
\arguments{
\item{configPath}{Path to global configuration path}

\item{dashboardConfig}{Dashboard configuration options (see create dashboard configuration file)}
}
\description{
Create stand-alone dashboaard database from reward data
}
\details{
Creates a configuration file for a Reward dashboard Shiny application to allow investigation in to
a subset of exposures or outcomes
}
