% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardCreator.R
\name{createDashboardDatabase}
\alias{createDashboardDatabase}
\title{Create stand-alone dashboaard database reward data}
\usage{
createDashboardDatabase(
  configPath,
  dashboardConfigPath,
  targetConnectionDetails = NULL,
  resultDatabaseSchema = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{configPath}{Path to global configuration path}

\item{dashboardConfigPath}{Dashboard configuration path for config options (see create dashboard configuration file)}

\item{targetConnectionDetails}{If Exporting to external database, use these connection details}
}
\description{
Create stand-alone dashboaard database reward data
}
\details{
Creates and sqlite database a set of configuration parameters that can be used with the Reward
Shiny application.
}
