% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{createDashboardResults}
\alias{createDashboardResults}
\title{Create Dashboard Results (Run SCC Analyses)}
\usage{
createDashboardResults(config = config::get(), dashboard)
}
\arguments{
\item{config}{Configuration list, usually from `config::get()`, containing database and data source details.}

\item{dashboard}{Dashboard configuration object, as created by `getDashboardConfig()`.}
}
\description{
Runs self-controlled cohort analyses for all data sources defined in the configuration and stores the results for the specified dashboard.
}
\details{
Determines SCC analysis settings and runs the analyses for each configured datasource, passing the appropriate exposure/outcome/target cohorts, controls, and dashboard configuration.
}
\seealso{
[execSccAnalyses()], [getNegativeControlPairs()]
}
