% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{createDatabaseMetaInfo}
\alias{createDatabaseMetaInfo}
\title{Create Database Meta Information Table for Dashboard}
\usage{
createDatabaseMetaInfo(config = config::get(), dashboard)
}
\arguments{
\item{config}{Project config.}

\item{dashboard}{Dashboard config.}
}
\description{
Extracts source metadata for all configured datasources and creates a `database_meta_data` table in the dashboard’s results schema.
}
\seealso{
[DatabaseConnector::insertTable()]
}
