% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{createGlobalConfig}
\alias{createGlobalConfig}
\title{Create global config file}
\usage{
createGlobalConfig(
  configPath,
  keyring = NULL,
  overwrite = FALSE,
  testConnection = TRUE
)
}
\arguments{
\item{keyring}{Optional keyring::keyring for storing secrets}

\item{overwrite}{Overwite existing file (if it exists)}

\item{testConnection}{Attempt to connect to database and write to schemas needed for writing?}
}
\description{
Create configuration file for REWARD server
}
