% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVocabularySchema.R
\name{createGlobalSchema}
\alias{createGlobalSchema}
\title{Create Results Schema and Core Tables for Global Analysis}
\usage{
createGlobalSchema(config = config::get())
}
\arguments{
\item{config}{List. Configuration object including credentials and schema name, usually from `config::get()`.}
}
\value{
Invisible `NULL`. Used for its side effects.
}
\description{
Drops and recreates the results schema (for non-vocabulary data) and initializes core results tables for self-controlled cohort and cohort generation packages.
}
\details{
- Prompts interactively before dropping any tables; the entire schema will be dropped and rebuilt.
- After creating the schema, templates provided by `SelfControlledCohort` and `CohortGenerator` are used to create all necessary results tables.
- Uses the schema set as `config$resultsDatabaseSchema`.
}
\examples{
\dontrun{
createGlobalSchema()
}

}
\seealso{
[SelfControlledCohort::createResultsDataModel()], [CohortGenerator::createResultsDataModel()]
}
