% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{createNullDist}
\alias{createNullDist}
\title{Create s3 class that matches EmpiricalCalibration's null}
\usage{
createNullDist(mean, sd)
}
\description{
Create s3 class that matches EmpiricalCalibration's null
}
