% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVocabularySchema.R
\name{createVocabularySchema}
\alias{createVocabularySchema}
\title{Create and Load OMOP Vocabulary Tables in Results Schema}
\usage{
createVocabularySchema(config = config::get())
}
\arguments{
\item{config}{List. Configuration object including database credentials and schema names, usually from `config::get()`.}
}
\value{
Invisible `NULL`. Called for its side effects (recreates and loads tables).
}
\description{
Drops (if exists) and recreates the vocabulary schema in the results database. Then creates all required OMOP vocabulary tables with appropriate indexes and primary keys, and bulk uploads vocabulary data from local CSV files (assumed to be in the directory `"vocabulary_tables"`).
}
\details{
- The function prompts for interactive confirmation before dropping the existing schema; **all tables in the schema will be deleted.**
- All tables (`concept`, `concept_ancestor`, `domain`, `drug_strength`, etc.) are (re)created from scratch with recommended OMOP fields and indexes.
- Vocabulary data is loaded from corresponding local CSVs (e.g. `vocabulary_tables/concept.csv`) using the `ResultModelManager::pyUploadCsv()` utility.
- Uses the schema name provided in `config$resultsVocabularyDatabaseSchema`.
}
\examples{
\dontrun{
createVocabularySchema()
}

}
\seealso{
[ResultModelManager::pyUploadCsv()]
}
