% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{DashboardDataModel}
\alias{DashboardDataModel}
\title{CEM Database Backend Class}
\description{
An interface to the common evidence model that uses works directly with a database schema
}
\section{Super class}{
\code{\link[Reward:RewardDataModel]{Reward::RewardDataModel}} -> \code{DashboardDataModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardDataModel$new()}}
\item \href{#method-getOutcomeCohorts}{\code{DashboardDataModel$getOutcomeCohorts()}}
\item \href{#method-getExposureCohorts}{\code{DashboardDataModel$getExposureCohorts()}}
\item \href{#method-getFilteredTableResultsQuery}{\code{DashboardDataModel$getFilteredTableResultsQuery()}}
\item \href{#method-getFilteredTableResults}{\code{DashboardDataModel$getFilteredTableResults()}}
\item \href{#method-getFilteredTableResultsCount}{\code{DashboardDataModel$getFilteredTableResultsCount()}}
\item \href{#method-getMetaAnalysisTable}{\code{DashboardDataModel$getMetaAnalysisTable()}}
\item \href{#method-getForestPlotTable}{\code{DashboardDataModel$getForestPlotTable()}}
\item \href{#method-getSummaryStats}{\code{DashboardDataModel$getSummaryStats()}}
\item \href{#method-getTimeToOutcomeStats}{\code{DashboardDataModel$getTimeToOutcomeStats()}}
\item \href{#method-getTimeOnTreatmentStats}{\code{DashboardDataModel$getTimeOnTreatmentStats()}}
\item \href{#method-getNegativeControlSccResults}{\code{DashboardDataModel$getNegativeControlSccResults()}}
\item \href{#method-clone}{\code{DashboardDataModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="countQuery">}\href{../../Reward/html/RewardDataModel.html#method-countQuery}{\code{Reward::RewardDataModel$countQuery()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="finalize">}\href{../../Reward/html/RewardDataModel.html#method-finalize}{\code{Reward::RewardDataModel$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getAnalysisSettings">}\href{../../Reward/html/RewardDataModel.html#method-getAnalysisSettings}{\code{Reward::RewardDataModel$getAnalysisSettings()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCemConnection">}\href{../../Reward/html/RewardDataModel.html#method-getCemConnection}{\code{Reward::RewardDataModel$getCemConnection()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohort">}\href{../../Reward/html/RewardDataModel.html#method-getCohort}{\code{Reward::RewardDataModel$getCohort()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohortConceptSet">}\href{../../Reward/html/RewardDataModel.html#method-getCohortConceptSet}{\code{Reward::RewardDataModel$getCohortConceptSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohortStats">}\href{../../Reward/html/RewardDataModel.html#method-getCohortStats}{\code{Reward::RewardDataModel$getCohortStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getDataSources">}\href{../../Reward/html/RewardDataModel.html#method-getDataSources}{\code{Reward::RewardDataModel$getDataSources()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getExposureCohortConceptSets">}\href{../../Reward/html/RewardDataModel.html#method-getExposureCohortConceptSets}{\code{Reward::RewardDataModel$getExposureCohortConceptSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getExposureCohortDefinitionSet">}\href{../../Reward/html/RewardDataModel.html#method-getExposureCohortDefinitionSet}{\code{Reward::RewardDataModel$getExposureCohortDefinitionSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getNegativeControlConditions">}\href{../../Reward/html/RewardDataModel.html#method-getNegativeControlConditions}{\code{Reward::RewardDataModel$getNegativeControlConditions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getNegativeControlExposures">}\href{../../Reward/html/RewardDataModel.html#method-getNegativeControlExposures}{\code{Reward::RewardDataModel$getNegativeControlExposures()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getOutcomeCohortConceptSets">}\href{../../Reward/html/RewardDataModel.html#method-getOutcomeCohortConceptSets}{\code{Reward::RewardDataModel$getOutcomeCohortConceptSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getOutcomeCohortDefinitionSet">}\href{../../Reward/html/RewardDataModel.html#method-getOutcomeCohortDefinitionSet}{\code{Reward::RewardDataModel$getOutcomeCohortDefinitionSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="queryDb">}\href{../../Reward/html/RewardDataModel.html#method-queryDb}{\code{Reward::RewardDataModel$queryDb()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$new(
  dashboardConfigPath = NULL,
  connectionDetails,
  cemConnectionDetails = NULL,
  resultDatabaseSchema = NULL,
  usePooledConnection = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dashboardConfigPath}}{Path to dashboart yaml file}

\item{\code{connectionDetails}}{Database Connector connection detils}

\item{\code{cemConnectionDetails}}{Cem Connection details}

\item{\code{resultDatabaseSchema}}{Results database schema}

\item{\code{usePooledConnection}}{Use a poooled database connection
Get Outcome Cohorts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOutcomeCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-getOutcomeCohorts}{}}}
\subsection{Method \code{getOutcomeCohorts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getOutcomeCohorts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame of cohorts
Get Exposure Cohorts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExposureCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-getExposureCohorts}{}}}
\subsection{Method \code{getExposureCohorts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getExposureCohorts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame of cohorts
Shiny Dashboard - main query
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFilteredTableResultsQuery"></a>}}
\if{latex}{\out{\hypertarget{method-getFilteredTableResultsQuery}{}}}
\subsection{Method \code{getFilteredTableResultsQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResultsQuery(
  benefitThreshold = 0.5,
  riskThreshold = 2,
  pValueCut = 0.05,
  requiredBenefitSources = NULL,
  filterByMeta = FALSE,
  outcomeCohortTypes = c(0, 1, 2, 3),
  calibrated = TRUE,
  benefitCount = 1,
  riskCount = 0,
  targetCohorts = NULL,
  outcomeCohorts = NULL,
  exposureClasses = NULL,
  orderByCol = NULL,
  ascending = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{benefitThreshold}}{thereshold to consider a benefit}

\item{\code{riskThreshold}}{threshold to consider effect estimate a risk}

\item{\code{pValueCut}}{pvalue hacking}

\item{\code{requiredBenefitSources}}{required sources to be in results}

\item{\code{filterByMeta}}{filter by meta analysis RR?}

\item{\code{outcomeCohortTypes}}{outcome cohorts to filter by type of}

\item{\code{calibrated}}{calibrated result}

\item{\code{benefitCount}}{minimum number of benefits found}

\item{\code{riskCount}}{number of accerptable risks}

\item{\code{targetCohorts}}{target cohort ids}

\item{\code{outcomeCohorts}}{outcome cohort ids}

\item{\code{exposureClasses}}{exposureClasses filter}

\item{\code{orderByCol}}{Order by which column?}

\item{\code{ascending}}{ascending order?}

\item{\code{limit}}{Row limit}

\item{\code{offset}}{Row Offset
Main table query in dashboard}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFilteredTableResults"></a>}}
\if{latex}{\out{\hypertarget{method-getFilteredTableResults}{}}}
\subsection{Method \code{getFilteredTableResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResults(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFilteredTableResultsCount"></a>}}
\if{latex}{\out{\hypertarget{method-getFilteredTableResultsCount}{}}}
\subsection{Method \code{getFilteredTableResultsCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResultsCount(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetaAnalysisTable"></a>}}
\if{latex}{\out{\hypertarget{method-getMetaAnalysisTable}{}}}
\subsection{Method \code{getMetaAnalysisTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getMetaAnalysisTable(exposureId, outcomeId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposureId}}{exposure cohort id}

\item{\code{outcomeId}}{outcome Cohort id
Get Forest plot table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getForestPlotTable"></a>}}
\if{latex}{\out{\hypertarget{method-getForestPlotTable}{}}}
\subsection{Method \code{getForestPlotTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getForestPlotTable(exposureId, outcomeId, calibrated)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposureId}}{exposure Id}

\item{\code{outcomeId}}{outcome Id}

\item{\code{calibrated}}{get calibrated results?
Get Summary Statistics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSummaryStats"></a>}}
\if{latex}{\out{\hypertarget{method-getSummaryStats}{}}}
\subsection{Method \code{getSummaryStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getSummaryStats(
  statType,
  exposureId,
  outcomeId,
  sourceIds = NULL,
  tableName = "scc_stat",
  analysisId = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statType}}{statistic type}

\item{\code{exposureId}}{exposure cohort id}

\item{\code{outcomeId}}{outcome cohort id}

\item{\code{sourceIds}}{cohort source ids}

\item{\code{tableName}}{table name}

\item{\code{analysisId}}{Analysis setting id
getTimeOnTreatmentStats}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTimeToOutcomeStats"></a>}}
\if{latex}{\out{\hypertarget{method-getTimeToOutcomeStats}{}}}
\subsection{Method \code{getTimeToOutcomeStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getTimeToOutcomeStats(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{getTimeOnTreatmentStats}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTimeOnTreatmentStats"></a>}}
\if{latex}{\out{\hypertarget{method-getTimeOnTreatmentStats}{}}}
\subsection{Method \code{getTimeOnTreatmentStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getTimeOnTreatmentStats(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Get Negative controls Scc results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNegativeControlSccResults"></a>}}
\if{latex}{\out{\hypertarget{method-getNegativeControlSccResults}{}}}
\subsection{Method \code{getNegativeControlSccResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getNegativeControlSccResults(
  cohortDefinitionId,
  isExposure,
  outcomeType = NULL,
  conceptSet = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionId}}{cohort id}

\item{\code{isExposure}}{is exposure or outcome?}

\item{\code{outcomeType}}{Outcome Type}

\item{\code{conceptSet}}{consept set data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
