% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{deployDashboard}
\alias{deployDashboard}
\title{Deploy a Dashboard to RStudio Connect or Shiny Server}
\usage{
deployDashboard(dashboardName, config = config::get())
}
\arguments{
\item{dashboardName}{Character. Dashboard name.}

\item{config}{Project config.}
}
\description{
Copies all necessary files, writes deployment environment variables, and deploys the dashboard app to the selected app directory using `rsconnect::deployApp`.
}
\seealso{
[rsconnect::deployApp()]
}
