% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{.authWebApi}
\alias{.authWebApi}
\title{Cross platform aut for WebApi}
\usage{
.authWebApi(config = config::get())
}
\arguments{
\item{config}{reward config}
}
\description{
Cross platform aut for WebApi
}
