% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{executePostProcessing}
\alias{executePostProcessing}
\title{Execute Workflow for post processing data}
\usage{
executePostProcessing(configPath)
}
\description{
Assumes results have been imported from CDM executions
}
