% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportReferences.R
\name{exportReferenceTables}
\alias{exportReferenceTables}
\title{Export Reference tables}
\usage{
exportReferenceTables(
  config,
  connection = NULL,
  exportPath = tempdir(),
  exportZipFile = "reward-references.zip"
)
}
\arguments{
\item{config}{global reward config}

\item{exportPath}{path to export files to before zipping}

\item{exportZipPath}{resulting zip files}
}
\description{
Takes created reference tables (cohort definitions) from central rewardb and exports them to a zipped csv file
}
