% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CemConnector.R
\name{getCemConnection}
\alias{getCemConnection}
\title{Get Connection to CEM API}
\usage{
getCemConnection(config = config::get())
}
\arguments{
\item{config}{assure configuration}
}
\description{
used for exploring evidence
}
\concept{{NegativeControls}}
