% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardConfig.R
\name{getDashboardConfig}
\alias{getDashboardConfig}
\title{Get dashboard configuration}
\usage{
getDashboardConfig(dashboardName)
}
\arguments{
\item{dashboardName}{config file name}
}
\description{
Get dashboard configuration
}
