% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{getDashboardDataModel}
\alias{getDashboardDataModel}
\title{Get Dashboard Data Model}
\usage{
getDashboardDataModel(dashboardName, config = config::get())
}
\arguments{
\item{dashboardName}{Character. Name of the dashboard.}

\item{config}{Configuration list.}
}
\value{
An `SccDataModel` object, ready for dashboard data operations.
}
\description{
Retrieves an SccDataModel instance for a given dashboard, enabling access to results and summary statistics for that dashboard.
}
\seealso{
[SelfControlledCohort:::SccDataModel]
}
