% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinitionSet.R
\name{getMockCohortDefinitionSet}
\alias{getMockCohortDefinitionSet}
\title{Get Mock Cohort definition set for subset of reward cohorts.}
\usage{
getMockCohortDefinitionSet(rewardConfig, cohortIds)
}
\arguments{
\item{rewardConfig}{Reward configuration}

\item{cohortIds}{Cohort Identifiers}
}
\description{
Useful in other analysis steps with ohdsi packages
}
