% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CemConnector.R
\name{getNegativeControlPairs}
\alias{getNegativeControlPairs}
\title{Get negative control pairs for a given dashboard config}
\usage{
getNegativeControlPairs(config = config::get(), dashboard)
}
\arguments{
\item{config}{assure configuration}
}
\description{
used for exploring evidence
}
\seealso{
Other {NegativeControls}: 
\code{\link{getCemConnection}()}
}
\concept{{NegativeControls}}
