% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getNullDist}
\alias{getNullDist}
\title{Get Null Distribution from EmpiricalCalibration Package}
\usage{
getNullDist(negatives)
}
\description{
Get Null Distribution from EmpiricalCalibration Package
}
