% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{getResultsConnectionDetails}
\alias{getResultsConnectionDetails}
\title{Get Results Connection Details}
\usage{
getResultsConnectionDetails(config = config::get())
}
\arguments{
\item{config}{List. Configuration object containing a results connection details list, typically from `config::get()`.}
}
\value{
A `DatabaseConnector::connectionDetails` object for use with functions requiring DB connections.
}
\description{
Returns a set of connection details suitable for use with DatabaseConnector, based on the configuration’s `resultsConnectionDetails`.
}
\seealso{
[DatabaseConnector::createConnectionDetails()]
}
