% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportResults.R
\name{importResultsFromS3}
\alias{importResultsFromS3}
\title{Import Results from CDM}
\usage{
importResultsFromS3(
  config,
  cdmManifestPath,
  connection = NULL,
  cleanup = TRUE,
  computeStatsTables = TRUE,
  numberOfThreads = 4
)
}
\arguments{
\item{config}{Reward configuration object}

\item{connection}{DatabaseConnector connection to reward databases}

\item{cleanup}{Remove extracted csv files after inserting results}

\item{numberOfThreads}{Number of upload jobs to pull and uploads}

\item{cdmManifest}{Path to cdm manifest json including location of s3 objects}

\item{computeAggregateTables}{compute tables that count aggregated stats across different data sources}
}
\description{
Verify and import results exported from CDMs into S3 bucket(s)

It is strongly advised that you use this utility on an EC2 box or elsewhere within the AWS cloud

Requires aws.s3 package and configured environment variables:
AWS_SSE_TYPE="AES256"
AWS_BUCKET_NAME
AWS_OBJECT_KEY
AWS_ACCESS_KEY_ID
AWS_SECRET_ACCESS_KEY
AWS_DEFAULT_REGION
}
