% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{importRewardCohorts}
\alias{importRewardCohorts}
\title{Use this function to refresh the library of cohorts in the project.
If any cohorts have changed since the last time this was run, they should redownload from atlas and re-generate}
\usage{
importRewardCohorts(config = config::get(), recache = FALSE)
}
\description{
Use this function to refresh the library of cohorts in the project.
If any cohorts have changed since the last time this was run, they should redownload from atlas and re-generate
}
