% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{launchCalibrationExplorer}
\alias{launchCalibrationExplorer}
\title{Launch shiny dashboard for exploration of calibration settings}
\usage{
launchCalibrationExplorer(configPath, usePooledConnection = TRUE)
}
\arguments{
\item{usedPooledConnection}{(Optional) Use a pooled database connection (used in multi-user environments).}
}
\description{
Shiny utility to expolore the results data set without biasing an investigator.
This tool allows the exploration of negative controls for a given exposure or outcome.
For this purpose an investigator can:
 * Check if negative controls for a given outcome or exposure exist without manual selection
 * View a calibration `fan plot` to examine the empirical null distribution
 * See the expected systematic error for a given exposure or outcome on each available data set
 * See the total number of results that would be included in a final analysis for the selected exposure or outcome
}
