% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-Dashboard.R
\name{launchDashboard}
\alias{launchDashboard}
\title{Launch the REWARD Shiny app dashboard}
\usage{
launchDashboard(connectionDetails, dashboardConfig)
}
\arguments{
\item{connectionDetails}{databaseConnector connection details}

\item{dashboardConfig}{see createDashboardConfig}
}
\description{
Launches a Shiny app for a given configuration file
}
