% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardCreatorUi.R
\name{launchDashboardCreator}
\alias{launchDashboardCreator}
\title{Shiny app to create stand-alone dashboaard database from reward data}
\usage{
launchDashboardCreator(configPath, usePooledConnection = TRUE)
}
\arguments{
\item{configPath}{Path to Reward global configuration file}
}
\description{
Shiny app to create stand-alone dashboaard database from reward data
}
\details{
Creates a configuration file for a Reward dashboard Shiny application to allow investigation in to
a subset of exposures or outcomes
}
