% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{launchShiny}
\alias{launchShiny}
\title{Launch the Shiny Dashboard Application}
\usage{
launchShiny(dashboardName, config = config::get())
}
\arguments{
\item{dashboardName}{Character. Name of dashboard to launch.}

\item{config}{Project configuration object.}
}
\description{
Loads the SelfControlledCohort package and launches the dashboard Shiny app for the given dashboard name.
}
\seealso{
[SelfControlledCohort::launchDashboard()]
}
