% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{loadGlobalConfiguration}
\alias{loadGlobalConfiguration}
\title{Loads global config}
\usage{
loadGlobalConfiguration(globalConfigPath, keyring = NULL)
}
\arguments{
\item{globalConfigPath}{Path to global configuration yaml file}

\item{keyring}{Optional keyring::keyring for storing secrets}
}
\description{
Load reward global config yaml file
}
