% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegativeControls.R
\name{mapNegativeControlExposures}
\alias{mapNegativeControlExposures}
\title{Map Negative Controls}
\usage{
mapNegativeControlExposures(configPath)
}
\description{
Map negative controls to outcome and exposure cohorts using OHDSI/CemConnector package
}
