% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegativeControls.R
\name{mapNegativeControlOutcomes}
\alias{mapNegativeControlOutcomes}
\title{Map Negative Controls}
\usage{
mapNegativeControlOutcomes(configPath)
}
\description{
Map negative controls to outcome and exposure cohorts using OHDSI/CemConnector package
}
