% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseSchema.R
\name{migrateDatabase}
\alias{migrateDatabase}
\title{Migrate database model}
\usage{
migrateDatabase(configFilePath)
}
\arguments{
\item{configFilePath}{path to global reward config}
}
\description{
Migrate database model
}
