% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-OpenTargetsModule.R
\name{openTargetsSearchModule}
\alias{openTargetsSearchModule}
\title{Open targets to omop vocabulary filtering shiny module}
\usage{
openTargetsSearchModule(id, model, openTargetsDatabaseSchema)
}
\arguments{
\item{id}{namespace id}

\item{model}{scc data model}

\item{openTargetsDatabaseSchema}{database schema where open targets results live}
}
\description{
Hooks up the search for open targets.
This returns a set of reactive values that can be used to filter concepts in the omop CDM.
}
