% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-OpenTargetsModule.R
\name{openTargetsSearchUiBlock}
\alias{openTargetsSearchUiBlock}
\title{UI For open targets search module}
\usage{
openTargetsSearchUiBlock(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Adds ui elements
}
