% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportResults.R
\name{registerCdm}
\alias{registerCdm}
\title{Register a CDM or return sourceId of existing cdm}
\usage{
registerCdm(config, connection, cdmInfoFile)
}
\arguments{
\item{config}{Reward config}

\item{connection}{connection to Reward db instance}

\item{cdmInfoFile}{path to configruations}
}
\description{
This function adds a CDM to reward - however, the config file is not fully required.
Only the meta-data fields `database` and `name` are required.

When a cdm is registered the database field is a unique idenitifier.
}
