% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reward.R
\docType{package}
\name{Reward-package}
\alias{Reward}
\alias{Reward-package}
\title{Reward: REal World Assessment and Research of Drug (REWARD) performance}
\description{
REWARD base package for managing data and constructing studies.
}
\keyword{internal}
