% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{RewardDataModel}
\alias{RewardDataModel}
\title{CEM Database Backend Class}
\description{
An interface to the common evidence model that uses works directly with a database schema
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{DatabaseConnector::connection instance}

\item{\code{vocabularySchema}}{OMOP vocabulary schema (must include concept and concept ancestor tables)}

\item{\code{resultsSchema}}{schema containing reward references and results}

\item{\code{config}}{Reward configuration s3 object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RewardDataModel-new}{\code{RewardDataModel$new()}}
\item \href{#method-RewardDataModel-finalize}{\code{RewardDataModel$finalize()}}
\item \href{#method-RewardDataModel-getCemConnection}{\code{RewardDataModel$getCemConnection()}}
\item \href{#method-RewardDataModel-getDataSources}{\code{RewardDataModel$getDataSources()}}
\item \href{#method-RewardDataModel-getOutcomeCohortDefinitionSet}{\code{RewardDataModel$getOutcomeCohortDefinitionSet()}}
\item \href{#method-RewardDataModel-getExposureCohortDefinitionSet}{\code{RewardDataModel$getExposureCohortDefinitionSet()}}
\item \href{#method-RewardDataModel-getExposureCohortConceptSets}{\code{RewardDataModel$getExposureCohortConceptSets()}}
\item \href{#method-RewardDataModel-getOutcomeCohortConceptSets}{\code{RewardDataModel$getOutcomeCohortConceptSets()}}
\item \href{#method-RewardDataModel-getCohortConceptSet}{\code{RewardDataModel$getCohortConceptSet()}}
\item \href{#method-RewardDataModel-getCohort}{\code{RewardDataModel$getCohort()}}
\item \href{#method-RewardDataModel-getAnalysisSettings}{\code{RewardDataModel$getAnalysisSettings()}}
\item \href{#method-RewardDataModel-getCohortStats}{\code{RewardDataModel$getCohortStats()}}
\item \href{#method-RewardDataModel-getNegativeControlSccResults}{\code{RewardDataModel$getNegativeControlSccResults()}}
\item \href{#method-RewardDataModel-clone}{\code{RewardDataModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-new}{}}}
\subsection{Method \code{new()}}{
initialize backend object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$new(configPath, keyring = NULL, usePooledConnection = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configPath}}{Reward configuration yaml path}

\item{\code{keyring}}{optional keyring::keyring object}

\item{\code{usePooledConnection}}{Used a pooled connection object rather than a database connector object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getCemConnection"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getCemConnection}{}}}
\subsection{Method \code{getCemConnection()}}{
Get cem connector api connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getCemConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getDataSources"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getDataSources}{}}}
\subsection{Method \code{getDataSources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getDataSources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getOutcomeCohortDefinitionSet"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getOutcomeCohortDefinitionSet}{}}}
\subsection{Method \code{getOutcomeCohortDefinitionSet()}}{
Get outcome cohort definition set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getOutcomeCohortDefinitionSet(cohortIds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{numeric vector of cohort ids or null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getExposureCohortDefinitionSet"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getExposureCohortDefinitionSet}{}}}
\subsection{Method \code{getExposureCohortDefinitionSet()}}{
Get exposure cohort definition set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getExposureCohortDefinitionSet(cohortIds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{numeric vector of cohort ids or null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getExposureCohortConceptSets"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getExposureCohortConceptSets}{}}}
\subsection{Method \code{getExposureCohortConceptSets()}}{
Get expsoure cohort concept sets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getExposureCohortConceptSets(cohortIds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{numeric vector of cohort ids or null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getOutcomeCohortConceptSets"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getOutcomeCohortConceptSets}{}}}
\subsection{Method \code{getOutcomeCohortConceptSets()}}{
Get outcome cohort concept sets for all cohorts specified
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getOutcomeCohortConceptSets(cohortIds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{numeric vector of cohort ids or null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getCohortConceptSet"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getCohortConceptSet}{}}}
\subsection{Method \code{getCohortConceptSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getCohortConceptSet(cohortDefinitionId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getCohort"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getCohort}{}}}
\subsection{Method \code{getCohort()}}{
Get getCohort data for one cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getCohort(cohortDefinitionId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionId}}{cohort identifier (not null, integer)
Decode raw base 64}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getAnalysisSettings"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getAnalysisSettings}{}}}
\subsection{Method \code{getAnalysisSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getAnalysisSettings(decode = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getCohortStats"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getCohortStats}{}}}
\subsection{Method \code{getCohortStats()}}{
Get getCohortStats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getCohortStats(cohortDefinitionId, isExposure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionId}}{cohort identifier (not null, integer)}

\item{\code{isExposure}}{Logical exposure cohort or not

Returns SCC results for a cohort id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-getNegativeControlSccResults"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-getNegativeControlSccResults}{}}}
\subsection{Method \code{getNegativeControlSccResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$getNegativeControlSccResults(
  cohortDefinitionId,
  isExposure,
  outcomeType = NULL,
  conceptSet = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RewardDataModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RewardDataModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RewardDataModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
