% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward-Dashboard.R
\name{rewardModule}
\alias{rewardModule}
\title{Dashboard instance}
\usage{
rewardModule(id = "Reward", model)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}
}
\description{
Requires a server appConfig instance to be loaded in environment see scoping of launchDashboard
This can be obtained with rewardb::loadappConfig(...)
}
