% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward-DashboardUi.R
\name{rewardUi}
\alias{rewardUi}
\title{Ui for rewardb dashboard}
\usage{
rewardUi(id = "Reward", appConfig)
}
\arguments{
\item{request}{shiny request object}
}
\description{
Ui for rewardb dashboard
}
