% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportReferences.R
\name{saveAtlasCohortRefs}
\alias{saveAtlasCohortRefs}
\title{Save atlas cohort references.}
\usage{
saveAtlasCohortRefs(config, connection, exportPath)
}
\arguments{
\item{config}{RewardConfig instance}

\item{connection}{DatabaseConnector connection}

\item{path}{Path to export data to}
}
\description{
Creates a cohort definition set that can be used in other packages
}
