% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-DashboardUi.R
\name{sccUi}
\alias{sccUi}
\title{Ui for rewardb dashboard}
\usage{
sccUi(id = "scc-module", dashboardConfig)
}
\arguments{
\item{request}{shiny request object}
}
\description{
Ui for rewardb dashboard
}
