% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{uploadConceptSets}
\alias{uploadConceptSets}
\title{Upload Extracted Concept Sets to Database}
\usage{
uploadConceptSets(config = config::get())
}
\arguments{
\item{config}{List. Configuration object; must specify export and database schema paths. Usually from `config::get()`.}
}
\value{
Invisible `NULL`. Used for its side effects (populating results tables in the database).
}
\description{
Creates tables and bulk uploads CSV files for concept sets, cohort-concept set mappings, and concept set names into results database.
Use after running `extractConceptSets()` to transfer concept sets to a database backend.
}
\details{
Drops (deletes) and recreates three results tables:
- `cg_concept_set`
- `cg_concept_set_name`
- `cg_cohort_concept_set`
Then uploads their respective CSVs constructed by `extractConceptSets()`.
}
\examples{
\dontrun{
uploadConceptSets()
}

}
\seealso{
[extractConceptSets()]
}
