% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboards.R
\name{uploadDashboardData}
\alias{uploadDashboardData}
\title{Upload Dashboard Data to Results Schema}
\usage{
uploadDashboardData(config = config::get(), dashboard)
}
\arguments{
\item{config}{List. Project configuration from `config::get()`.}

\item{dashboard}{Dashboard config object holding database schema details.}
}
\description{
Uploads SelfControlledCohort and CohortGenerator results for all SCC analyses and all datsources to the dashboard’s results schema in the database.
}
\details{
Prompts before dropping and recreating the dashboard results schema. Then uploads analytic results for each SCC analysis for each datasource.
}
\seealso{
[SelfControlledCohort::uploadResults()]
}
