% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{validateConfigFile}
\alias{validateConfigFile}
\title{Validate a cdm configuration file}
\usage{
validateConfigFile(configPath, testConnection = TRUE, keyring = NULL)
}
\arguments{
\item{testConnection}{Attempt to connect to database and write to schemas needed for writing?}

\item{keyring}{Optional keyring::keyring for storing secrets}
}
\description{
Opens a file for editing that contains the default settings for a cdm
}
