% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationPlotModule.R
\name{calibrationPlotServer}
\alias{calibrationPlotServer}
\title{Calibration plot siny server module}
\usage{
calibrationPlotServer(id, model, selectedCohort)
}
\arguments{
\item{id}{shiny namespace should be consistent with UI}

\item{model}{RewardDataModel R6 class instance}

\item{selectedCohort}{cohort object reactive - returned list must contain: cohortDefinitionId, isExposure, selectedOutcomeType (reactive)}
}
\description{
Display calibration plots for reward cohorts based on data available in system
Also shows Expected Absolute Systematic Error
}
