% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardData.R
\name{createDashboardDatabase}
\alias{createDashboardDatabase}
\title{Create stand-alone dashboaard database from reward data}
\usage{
createDashboardDatabase(configPath, dashboardConfig, overwrite = FALSE)
}
\arguments{
\item{configPath}{Path to global configuration path}

\item{dashboardConfig}{Dashboard configuration options (see create dashboard configuration file)}
}
\description{
Create stand-alone dashboaard database from reward data
}
\details{
Creates and sqlite database from a set of configuration parameters that can be used with the Reward
Shiny application.
}
