% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{DashboardDataModel}
\alias{DashboardDataModel}
\title{CEM Database Backend Class}
\description{
An interface to the common evidence model that uses works directly with a database schema
}
\section{Super class}{
\code{\link[Reward:RewardDataModel]{Reward::RewardDataModel}} -> \code{DashboardDataModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DashboardDataModel-new}{\code{DashboardDataModel$new()}}
\item \href{#method-DashboardDataModel-getOutcomeCohorts}{\code{DashboardDataModel$getOutcomeCohorts()}}
\item \href{#method-DashboardDataModel-getExposureCohorts}{\code{DashboardDataModel$getExposureCohorts()}}
\item \href{#method-DashboardDataModel-getFilteredTableResultsQuery}{\code{DashboardDataModel$getFilteredTableResultsQuery()}}
\item \href{#method-DashboardDataModel-getFilteredTableResults}{\code{DashboardDataModel$getFilteredTableResults()}}
\item \href{#method-DashboardDataModel-getFilteredTableResultsCount}{\code{DashboardDataModel$getFilteredTableResultsCount()}}
\item \href{#method-DashboardDataModel-getMetaAnalysisTable}{\code{DashboardDataModel$getMetaAnalysisTable()}}
\item \href{#method-DashboardDataModel-getForestPlotTable}{\code{DashboardDataModel$getForestPlotTable()}}
\item \href{#method-DashboardDataModel-getSummaryStats}{\code{DashboardDataModel$getSummaryStats()}}
\item \href{#method-DashboardDataModel-getTimeToOutcomeStats}{\code{DashboardDataModel$getTimeToOutcomeStats()}}
\item \href{#method-DashboardDataModel-getTimeOnTreatmentStats}{\code{DashboardDataModel$getTimeOnTreatmentStats()}}
\item \href{#method-DashboardDataModel-getNegativeControlSccResults}{\code{DashboardDataModel$getNegativeControlSccResults()}}
\item \href{#method-DashboardDataModel-clone}{\code{DashboardDataModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="countQuery"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-countQuery'><code>Reward::RewardDataModel$countQuery()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="finalize"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-finalize'><code>Reward::RewardDataModel$finalize()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getAnalysisSettings"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getAnalysisSettings'><code>Reward::RewardDataModel$getAnalysisSettings()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCemConnection"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getCemConnection'><code>Reward::RewardDataModel$getCemConnection()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohort"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getCohort'><code>Reward::RewardDataModel$getCohort()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohortConceptSet"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getCohortConceptSet'><code>Reward::RewardDataModel$getCohortConceptSet()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getCohortStats"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getCohortStats'><code>Reward::RewardDataModel$getCohortStats()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getDataSources"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getDataSources'><code>Reward::RewardDataModel$getDataSources()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getExposureCohortConceptSets"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getExposureCohortConceptSets'><code>Reward::RewardDataModel$getExposureCohortConceptSets()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getExposureCohortDefinitionSet"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getExposureCohortDefinitionSet'><code>Reward::RewardDataModel$getExposureCohortDefinitionSet()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getNegativeControlConditions"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getNegativeControlConditions'><code>Reward::RewardDataModel$getNegativeControlConditions()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getNegativeControlExposures"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getNegativeControlExposures'><code>Reward::RewardDataModel$getNegativeControlExposures()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getOutcomeCohortConceptSets"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getOutcomeCohortConceptSets'><code>Reward::RewardDataModel$getOutcomeCohortConceptSets()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="getOutcomeCohortDefinitionSet"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-getOutcomeCohortDefinitionSet'><code>Reward::RewardDataModel$getOutcomeCohortDefinitionSet()</code></a></li>
<li><span class="pkg-link" data-pkg="Reward" data-topic="RewardDataModel" data-id="queryDb"><a href='../../Reward/html/RewardDataModel.html#method-RewardDataModel-queryDb'><code>Reward::RewardDataModel$queryDb()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$new(
  dashboardConfigPath = NULL,
  connectionDetails,
  cemConnectionDetails = NULL,
  resultDatabaseSchema = NULL,
  usePooledConnection = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dashboardConfigPath}}{Path to dashboart yaml file}

\item{\code{connectionDetails}}{Database Connector connection detils}

\item{\code{cemConnectionDetails}}{Cem Connection details}

\item{\code{resultDatabaseSchema}}{Results database schema}

\item{\code{usePooledConnection}}{Use a poooled database connection
Get Outcome Cohorts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getOutcomeCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getOutcomeCohorts}{}}}
\subsection{Method \code{getOutcomeCohorts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getOutcomeCohorts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame of cohorts
Get Exposure Cohorts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getExposureCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getExposureCohorts}{}}}
\subsection{Method \code{getExposureCohorts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getExposureCohorts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame of cohorts
Shiny Dashboard - main query
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getFilteredTableResultsQuery"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getFilteredTableResultsQuery}{}}}
\subsection{Method \code{getFilteredTableResultsQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResultsQuery(
  benefitThreshold = 0.5,
  riskThreshold = 2,
  pValueCut = 0.05,
  requiredBenefitSources = NULL,
  filterByMeta = FALSE,
  outcomeCohortTypes = c(0, 1, 2, 3),
  calibrated = TRUE,
  benefitCount = 1,
  riskCount = 0,
  targetCohorts = NULL,
  outcomeCohorts = NULL,
  exposureClasses = NULL,
  orderByCol = NULL,
  ascending = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{benefitThreshold}}{thereshold to consider a benefit}

\item{\code{riskThreshold}}{threshold to consider effect estimate a risk}

\item{\code{pValueCut}}{pvalue hacking}

\item{\code{requiredBenefitSources}}{required sources to be in results}

\item{\code{filterByMeta}}{filter by meta analysis RR?}

\item{\code{outcomeCohortTypes}}{outcome cohorts to filter by type of}

\item{\code{calibrated}}{calibrated result}

\item{\code{benefitCount}}{minimum number of benefits found}

\item{\code{riskCount}}{number of accerptable risks}

\item{\code{targetCohorts}}{target cohort ids}

\item{\code{outcomeCohorts}}{outcome cohort ids}

\item{\code{exposureClasses}}{exposureClasses filter}

\item{\code{orderByCol}}{Order by which column?}

\item{\code{ascending}}{ascending order?}

\item{\code{limit}}{Row limit}

\item{\code{offset}}{Row Offset
Main table query in dashboard}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getFilteredTableResults"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getFilteredTableResults}{}}}
\subsection{Method \code{getFilteredTableResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResults(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getFilteredTableResultsCount"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getFilteredTableResultsCount}{}}}
\subsection{Method \code{getFilteredTableResultsCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getFilteredTableResultsCount(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getMetaAnalysisTable"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getMetaAnalysisTable}{}}}
\subsection{Method \code{getMetaAnalysisTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getMetaAnalysisTable(exposureId, outcomeId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposureId}}{exposure cohort id}

\item{\code{outcomeId}}{outcome Cohort id
Get Forest plot table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getForestPlotTable"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getForestPlotTable}{}}}
\subsection{Method \code{getForestPlotTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getForestPlotTable(exposureId, outcomeId, calibrated)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposureId}}{exposure Id}

\item{\code{outcomeId}}{outcome Id}

\item{\code{calibrated}}{get calibrated results?
Get Summary Statistics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getSummaryStats"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getSummaryStats}{}}}
\subsection{Method \code{getSummaryStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getSummaryStats(
  statType,
  exposureId,
  outcomeId,
  sourceIds = NULL,
  tableName = "scc_stat",
  analysisId = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statType}}{statistic type}

\item{\code{exposureId}}{exposure cohort id}

\item{\code{outcomeId}}{outcome cohort id}

\item{\code{sourceIds}}{cohort source ids}

\item{\code{tableName}}{table name}

\item{\code{analysisId}}{Analysis setting id
getTimeOnTreatmentStats}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getTimeToOutcomeStats"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getTimeToOutcomeStats}{}}}
\subsection{Method \code{getTimeToOutcomeStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getTimeToOutcomeStats(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{getTimeOnTreatmentStats}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getTimeOnTreatmentStats"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getTimeOnTreatmentStats}{}}}
\subsection{Method \code{getTimeOnTreatmentStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getTimeOnTreatmentStats(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Get Negative controls Scc results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-getNegativeControlSccResults"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-getNegativeControlSccResults}{}}}
\subsection{Method \code{getNegativeControlSccResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$getNegativeControlSccResults(
  cohortDefinitionId,
  isExposure,
  outcomeType = NULL,
  conceptSet = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionId}}{cohort id}

\item{\code{isExposure}}{is exposure or outcome?}

\item{\code{outcomeType}}{Outcome Type}

\item{\code{conceptSet}}{consept set data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DashboardDataModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DashboardDataModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardDataModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
