% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward-Dashboard.R
\name{launchDashboard}
\alias{launchDashboard}
\title{Launch the REWARD Shiny app dashboard}
\usage{
launchDashboard(
  dashboardConfigPath,
  configPath = NULL,
  connectionDetails = NULL,
  resultDatabaseSchema = NULL
)
}
\arguments{
\item{appConfigPath}{path to configuration file. This is loaded in to the local environment with the appConfig variable}
}
\description{
Launches a Shiny app for a given configuration file
}
