% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportResults.R
\name{registerCdm}
\alias{registerCdm}
\title{Register a CDM or return sourceId of existing cdm}
\usage{
registerCdm(config, connection, cdmInfoFile)
}
\description{
Register a CDM or return sourceId of existing cdm
}
